SHIFT: axolotl: add initial config
The SHIFT6mq (axolotl) is an SDM845-based phone with a WCD934x codec.
It has the following audio devices:
- Headset
- Speaker
- Earpiece
- Headset microphone
- Back microphone
- Bottom microphone

It exposes 6 different MultiMedia* devices for routing the audio over
SLIMBUS and als a VoiceMMode1 for routing audio from and to the modem
for making phone calls.
To cover its use cases, a default HiFi use case is provided to play
audio from music, videos, etc. and a VoiceCall use case to make phone
calls.