Add q6voice routes in HiFi conf too
In OP6, there seems to be a race condition where q6voiced opens the
voicemmode device before the routes are configured, leading to:

[   85.540643]  VoiceMMode1: ASoC: no backend DAIs enabled for VoiceMMode1
[   85.540678]  VoiceMMode1: ASoC: error at dpcm_fe_dai_prepare on VoiceMMode1: -22
[   85.540724]  VoiceMMode1: ASoC: no backend DAIs enabled for VoiceMMode1
[   85.540733]  VoiceMMode1: ASoC: error at dpcm_fe_dai_prepare on VoiceMMode1: -22

So, add them to HiFi conf too, so the mixer confs are in place even if
voice profile get switched to slightly later. To be safe, do the same
workaround for OP6T and Poco F1 too.